+++ spec/spec_helper.rb
@@ -1,18 +1,30 @@
+require "bundler/setup"
 require 'rdf/spec/enumerable'
 require 'rdf/spec'
 require 'rdf/isomorphic'
 require 'rdf/ntriples'
 require 'rdf/turtle'
 require 'rdf/vocab'
-require 'simplecov'
-require 'coveralls'
+require 'amazing_print'
 
-SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new([
-  SimpleCov::Formatter::HTMLFormatter,
-  Coveralls::SimpleCov::Formatter
-])
-SimpleCov.start do
-  add_filter '/.bundle/'
+begin
+  require 'simplecov'
+  require 'simplecov-lcov'
+
+  SimpleCov::Formatter::LcovFormatter.config do |config|
+    #Coveralls is coverage by default/lcov. Send info results
+    config.report_with_single_file = true
+    config.single_report_path = 'coverage/lcov.info'
+  end
+
+  SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new([
+    SimpleCov::Formatter::HTMLFormatter,
+    SimpleCov::Formatter::LcovFormatter
+  ])
+  SimpleCov.start do
+    add_filter '/.bundle/'
+  end
+rescue LoadError
 end
 
 require 'spira'